Fix merge conflicts (#2)
* Update bindgen config to use `usize`

* Update macOS bindings

* Update Linux bindings

* Update Windows (x86_64) bindings

* Update Windows (x86) bindings

* Use `usize` instead of `u64`

* Use conditional function calling conventions

* Re-enable CI for Windows 32bit

* Fix CI matrix

* Add generated bindings for Apple M1

* On macOS Big Sur, homebrew installs to /opt/homebrew instead of /usr/local, adapt installation notes

* Feature is 'generate-bindings', not 'bindgen'. Change error message

* Add more paths to include directory, required when compilling and installing from source

* Formatting

* Compilation notes: environment variable was changed to 'ORT_LIB_LOCATION' (from 'ONNXRUNTIME_INSTALL_DIR')

* Add entry to changelog

* Add macOS aarch64 to README

* Add note about LD_LIBRARY_PATH in README

* Add a VSCode config to run clippy and format on save

* Clippy: Allow upper case acronyms, names coming from specific ML models

* Fix clippy lint: Change 'Into' impls to 'From'

* Add ORT_LIB_LOCATION note in readme for 'ORT_STRATEGY=system'

* Update changelog following #70

* Update macos/aarch64 bindings following #70

* Formatting

* Upgrade 'ndarray' to 0.15 (from 0.13)

* Upgrade 'ureq' to 2.1 (from 1.5)

* Add changelog entries

* Clippy: Box the larse ureq::Error in OrtDownloadError

* Release v0.0.12

* Return an Err on shape mismatch instead of panic

This way we let the user of the library decide how to handle the issue.

+++ onnxruntime/src/environment.rs
@@ -323,7 +323,7 @@ mod tests {
 
         let initial_name = String::from("concurrent_environment_creation");
         let main_env = Environment::new(initial_name.clone(), LoggingLevel::Warning).unwrap();
-        let main_env_ptr = main_env.env_ptr() as u64;
+        let main_env_ptr = main_env.env_ptr() as usize;
 
         let children: Vec<_> = (0..10)
             .map(|t| {
@@ -337,13 +337,13 @@ mod tests {
                         .unwrap();
 
                     assert_eq!(env.name(), initial_name_cloned);
-                    assert_eq!(env.env_ptr() as u64, main_env_ptr);
+                    assert_eq!(env.env_ptr() as usize, main_env_ptr);
                 })
             })
             .collect();
 
         assert_eq!(main_env.name(), initial_name);
-        assert_eq!(main_env.env_ptr() as u64, main_env_ptr);
+        assert_eq!(main_env.env_ptr() as usize, main_env_ptr);
 
         let res: Vec<std::thread::Result<_>> =
             children.into_iter().map(|child| child.join()).collect();
